remove accel.teardown and move test to test_trainer.py
